rrpmbuild.pl: second attempt for hard links (unsuccessful so disabled)
For some reason digests does not match, also no-digest is no-go
(rpm(8) still complains digest mismatch so some 'ordering' may affect).

The cpio can be extracted with cpio(1); hard and symbolic links extracted
as expected.

One change that could have affected is that directory has now link count
of (only) one (it was two (2)). man 5 cpio says
   "Directories always have a  value  of at  least  two here."
maybe that is just for that one format, but that is not informed
differently later. man 5 cpio says also (for new ascii format):
   "hardlinked files are handled by setting the filesize to zero
    for each entry except the first one that appears in the archive."
but it is not so, last one has the content, others zero.

"Thanks for playing, try once more..."