rrpmbuild.pl: revert c7757a6aa (parent commit) -- sort order
In case of rpm < 4.14 and there are hard links in an rpm
package, in order the package to be installable, files ordered in
the rpm header tags needs to be so that "dir.file" is ordered
before "dir/file". depth-first ordering ('dir', contents of 'dir'
and then 'file.dir'), which parent commit did, does not work.

Order of files in cpio is affected accodingly, just that the order
does not matter there.