rrpmbuild.pl: rpm packages with hard links installable with rpm < 4.14
In order for rpm packages to be installable with rpm(8) versions prior
to 4.14 the files in header tags needs to be ordered like the full
paths are ordered in find path/ | LC_ALL sort (i.e. in ascii order,
where e.g 'dir.file' is listed before 'dir/file').

That is now done and rrpmbuild.pl is working. Some polishment to be
done there, but it is expected polished version reproduces exactly
the same rpm package when running ./test.sh 1